WIP. Added ban_pubkey_rpc for testing convenience.
