+++ mm2src/lp_ordermatch.rs
@@ -61,6 +61,7 @@ use crate::mm2::lp_swap::{calc_max_maker_vol, check_balance_for_maker_swap, chec
                           lp_atomic_locktime, run_maker_swap, run_taker_swap, AtomicLocktimeVersion,
                           CheckBalanceError, MakerSwap, RunMakerSwapInput, RunTakerSwapInput,
                           SwapConfirmationsSettings, TakerSwap};
+
 pub use best_orders::best_orders_rpc;
 pub use orderbook_depth::orderbook_depth_rpc;
 pub use orderbook_rpc::orderbook_rpc;
